Remove string code, redirect to sort.Strings()
After some research, string sorting is too dependent on the kinds of
strings you are sorting.  Trie-based BurstSort and 3-way quick sorts
are often good for real world data, but do not work as well on random
strings.  I’m going to consider strings out of scope for this package.