Allow up to twice the cutoff in breakeven test
Allow twice the cutoff length before failing the breakeven test.

On some platforms the default cutover values before falling back to
the standard library sort aren't quite optimal.  The test isn't so much
to ensure that the choice is perfect for everyone, but rather that it's
good enough.

I'm unsure if this actually matters enough to bother to make it a parameter. While x86_64 is fairly popular, I want zermelo to support arm as well, but maybe not enough to mess with the API.