Add support for generics
This commit represents the biggest change to `zermelo` in years as it
add support for go 1.18 generics.  This means most of the old
subpackages are now obsolete, along with the old `Sorter` and `Sort`.

The tests have been thrown out and rewritten, and for now there's no
longer any benchmarks. This will likely be cleaned up over time.

The intention here is to keep backwards compatability with the old
non-generic interface, but there may have been mistakes made here and
there. This branch (likely to be called 1.5.x) will strive to keep that
API, but there will be a v2 eventually that will drop all that and
thereby greatly simplify the public API.

For now this should be considered relatively rough code, compared to
the stability of zermelo up to now, but I've done what I can to test
it extensively. Any bug reports are greatly appreciated.